requirements.txt
flask==2.1.0
google-api-python-client==2.41.0
google-generativeai==0.3.0
python-dotenv==0.19.2